store: move Lister and ListOptions back into store.go
Issue: https://github.com/adamdecaf/cert-manage/issues/174
